src/honeygpt/requirements.txt
requests
pexpect
pydantic
pytest
python-dotenv
